+++ common/src/main/java/net/SpectrumFATM/BlackArchive.java
@@ -1,11 +1,11 @@
 package net.SpectrumFATM;
 
+import earth.terrarium.adastra.api.systems.GravityApi;
 import net.SpectrumFATM.black_archive.block.ModBlocks;
 import net.SpectrumFATM.black_archive.blockentity.ModConsoles;
 import net.SpectrumFATM.black_archive.blockentity.ModShells;
 import net.SpectrumFATM.black_archive.effects.ModEffects;
 import net.SpectrumFATM.black_archive.entity.ModEntities;
-import net.SpectrumFATM.black_archive.entity.client.ModModelLayers;
 import net.SpectrumFATM.black_archive.item.ModItems;
 import net.SpectrumFATM.black_archive.network.BlackArchiveNetworkHandler;
 import net.SpectrumFATM.black_archive.sound.ModSounds;
@@ -13,7 +13,6 @@ import net.SpectrumFATM.black_archive.tardis.control.ModControls;
 import net.SpectrumFATM.black_archive.tardis.upgrades.ModUpgrades;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import whocraft.tardis_refined.api.event.TardisClientEvents;
 
 public final class BlackArchive {
     public static final String MOD_ID = "black_archive";