+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Time fissures now have a chance of spawning Silurians and Weeping Angels, as well as the already added mobs.
 - Added time lock control.
 - Sonic pen.
+- Ad Astra compatability.
 
 ### Removed
 - Rani console, exterior and interior.